Rebuilt
Thankfully Ben required us to print the code on physical paper.  I found
a copy of the final client and server code, but I was missing the
"mftp.h" header file. I seem to have rebuilt it as needed. Another lucky
break that the old repository had everything in the mftpserv.c file
before I had moved it.